requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
stripe==2.60.0
python-dotenv==0.19.0